requirements.txt
streamlit>=1.35.0
google-generativeai>=0.7.0
pinecone-client>=3.0.0
sentence-transformers>=3.0.0
PyMuPDF>=1.24.0
python-docx>=1.1.0
python-dotenv>=1.0.0
